rename db_v2 to database
